chore(sample): improve symbol uploads
Fixes #73
